- Rename and fix wrong attributes - Update webpack config - Remove an unused file
