+++ src/icons/templates/summernote.css
