--- src/icons/templates/summernote.json
@@ -1,5 +1,0 @@
-{
-	"baseClass": "",
-	"classPrefix": "note-icon-",
-	"mixinPrefix": "note-icon-"
-}
