+++ src/js/lite/settings.js
@@ -2,7 +2,7 @@ import $ from 'jquery';
 import ui from './ui';
 import '../base/settings.js';
 
-import '../../less/summernote-lite.less';
+import '../../styles/summernote-lite.less';
 
 $.summernote = $.extend($.summernote, {
   ui: ui,
