+++ src/styles/lite-ui/toolbar.less
