+++ src/styles/lite-ui/variables.less
