+++ src/summernote-bs3.html
@@ -32,8 +32,9 @@
     <div class="container">
       <div style='float:right;'>
         <a href='bs4.html'>Bootstrap4</a>
+        /
         <a href='lite.html'>Lite</a>
-        -
+        /
         <a href="examples/">Example</a>
       </div>
       <h1>Summernote with Bootstrap 3 </h1>
