+++ src/summernote-bs4.html
@@ -27,8 +27,9 @@
 <div class="container">
     <div style='float:right;'>
       <a href='index.html'>Bootstrap 3</a> 
+      /
       <a href='lite.html'>Lite</a>
-      -
+      /
       <a href="examples/">Example</a>
     </div>  
   <h1>Summernote with Bootstrap 4</h1>
