+++ src/summernote-lite.html
@@ -23,8 +23,9 @@
 <body>
   <div style='float:right;'>
     <a href='index.html'>Bootstrap 3</a> 
+    /
     <a href='bs4.html'>Bootstrap 4</a>
-    -
+    /
     <a href="examples/">Example</a>
   </div>  
   <h1>Summernote standalone</h1>
